requirements.txt
boto3
fastjsonschema
openai>=1.23.6
pyyaml
